+++ src/main/java/com/wuwei/service/StudentService.java
@@ -15,9 +15,9 @@ public interface StudentService {
 
     public List<Student> getAllStudent();
 
-    public Student getStudentById(int id);
+    public Student getStudentById(Long id);
 
     public int updateStudent(Student student);
 
-    public int delStudentById(int id);
+    public int delStudentById(Long id);
 }
